# case-tracking/case-tracking-service-with-rust/openapi.yaml
openapi: 3.1.0
info:
  title: Case Tracking API (Loco edition)
  version: 0.1.0
  description: >
    JSON API for tracking the physical location of paper case-note folders
    across NHS hospital cabinets. The tracker owns no domain tables; every
    entity is proxied from one of five upstream Main-X-Services. See
    `spec/index.md` for the full specification. This document is the
    machine-readable form of `spec/api-contract.md` + `spec/routes.md`.
servers:
  - url: http://localhost:5150
    description: Local dev (JSON API). In dev the SvelteKit app proxies /api here.
...
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /api/workers:
    get:
      tags: [workers]
      parameters: [{ name: q, in: query, schema: { type: string } }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/WorkerList' } } }
          description: OK
        '503': { $ref: '#/components/responses/Upstream' }
  /api/workers/{id}:
    get:
      tags: [workers]
      summary: Worker detail — folders they moved + all their patients' folders.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/WorkerShow' } } }
          description: OK
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Upstream' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cts_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    NotFound:
      description: Resource not found.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Upstream:
      description: An upstream Main-X-Service is unreachable.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      properties: { error: { type: string } }
...
    Folder:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        patient_id: { type: string, format: uuid }
        nhs_number: { type: string }
        patient_name: { type: string }
        cabinet_id: { type: [string, 'null'], format: uuid }
        cabinet_label: { type: string }
        status: { type: string, enum: [in-cabinet, in-transit] }
        last_moved_at: { type: [string, 'null'], format: date-time }
        notes: { type: [string, 'null'] }
        volume_id: { type: [string, 'null'], format: uuid }
        volume_title: { type: [string, 'null'] }
...
    Worker:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        role: { type: [string, 'null'] }
    WorkerShow:
      type: object
      properties:
        worker: { $ref: '#/components/schemas/Worker' }
        moved_folders: { type: array, items: { $ref: '#/components/schemas/Folder' } }
        patient_folders: { type: array, items: { $ref: '#/components/schemas/Folder' } }
        moves: { type: array, items: { $ref: '#/components/schemas/Move' } }
    Move:
      type: object
      properties:
        id: { type: string, format: uuid }
        folder_id: { type: string, format: uuid }
        folder_title: { type: string }
        patient_id: { type: string, format: uuid }
        nhs_number: { type: string }
        patient_name: { type: string }
        from_cabinet_id: { type: [string, 'null'], format: uuid }
        from_cabinet_label: { type: string }
        to_cabinet_id: { type: [string, 'null'], format: uuid }
        to_cabinet_label: { type: string }
        worker_id: { type: [string, 'null'], format: uuid }
        moved_by: { type: string }
        worker_role: { type: [string, 'null'] }
        moved_at: { type: string, format: date-time }
        reason: { type: [string, 'null'] }
...
    WorkerList:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Worker' } }
        query: { type: string }
...